Add newBook form and model book
