+++ routes/books.routes.js
@@ -4,5 +4,6 @@ const router = express.Router();
 const booksController = require('../controllers/books.controller');
 
 router.get('/home', booksController.get_books);
+router.get('/newBook',booksController.get_new);
 
 module.exports = router;
