+++ views/includes/head.ejs
@@ -19,7 +19,7 @@
     
       <div id="navbarExampleTransparentExample" class="navbar-menu">
         <div class="navbar-start">
-          <a class="navbar-item" href="https://bulma.io/">
+          <a class="navbar-item" href="/books/home">
             Home
           </a>
         <div class="navbar-end">
@@ -31,7 +31,7 @@
                     </a>
               </p>
               <p class="control">
-                <a class="button is-primary" href=#>
+                <a class="button is-black" href=#>
                   <span>Log in</span>
                 </a>
               </p>
